// project/src/types/film-data.ts
import { FavoriteAction } from '../const';

export interface IFilmData {
  id: number;
  name: string;
  posterImage: string;
  previewImage: string;
  backgroundImage: string;
  backgroundColor: string;
  videoLink: string;
  previewVideoLink: string;
  description: string;
  rating: number;
  scoresCount: number;
  director: string;
  starring: string[];
  runTime: number;
  genre: string;
  released: number;
  isFavorite: boolean;
}

export interface IFavoriteRequest {
  filmId: number,
  status: FavoriteAction
}